+++ src/components/CardList.jsx
@@ -1,25 +1,16 @@
 import CardListItem from "./CardListItem";
-import NoPage from "./NoPage";
-import useSelectFolder from "../hooks/useSelectFolder";
 
-function CardList({ url }) {
-  const result = useSelectFolder({ url });
-  const links = result?.data;
+function CardList({ links }) {
+  console.log(links);
   return (
-    <>
-      {links && links.length === 0 ? (
-        <NoPage />
-      ) : (
-        <ul className="CardLinkList">
-          {links &&
-            links.map((link) => (
-              <li key={link.id}>
-                <CardListItem link={link} />
-              </li>
-            ))}
-        </ul>
-      )}
-    </>
+    <ul className="CardLinkList">
+      {links &&
+        links.map((link) => (
+          <li key={link.id}>
+            <CardListItem link={link} />
+          </li>
+        ))}
+    </ul>
   );
 }
